Cast gnm_float to double in printf.
* dialog-simulation.c (simulation_update_sensitivity_cb)
(update_results_view): Cast gnm_float to double in printf.

* dialog-workbook-attr.c (struct AttrState, get_entry_values,
cb_widget_changed): Make iteration_tolerance gnm_float instead of
double.

+++ src/dialogs/dialog-simulation.c
@@ -187,12 +187,13 @@ update_log (SimulationState *state, simulation_t *sim)
 			g_string_append_printf (buf, "%d", sim->n_output_vars);
 			break;
 		case 4:
-			g_string_append_printf (buf, "%.2g",
-						sim->end.tv_sec -
-						sim->start.tv_sec +
-						(sim->end.tv_usec -
-						 sim->start.tv_usec) /
-						(gnm_float) G_USEC_PER_SEC);
+			g_string_append_printf 
+				(buf, "%.2g",
+				 (double) (sim->end.tv_sec - 
+					   sim->start.tv_sec +
+					   (sim->end.tv_usec -
+					    sim->start.tv_usec) /
+					   (gnm_float) G_USEC_PER_SEC));
 			break;
 		case 5:
 			dao_append_date (buf);
@@ -238,9 +239,9 @@ update_results_view (simulation_t *sim)
 	for (i = 0; i < sim->n_vars; i++)
 		g_string_append_printf (buf, "%-20s %10g %10g %10g\n",
 				   sim->cellnames [i],
-				   sim->stats [results_sim_index]->min [i],
-				   sim->stats [results_sim_index]->mean [i],
-				   sim->stats [results_sim_index]->max [i]);
+				   (double) sim->stats [results_sim_index]->min [i],
+				   (double) sim->stats [results_sim_index]->mean [i],
+				   (double) sim->stats [results_sim_index]->max [i]);
 
 	gtk_text_buffer_set_text (results_buffer, buf->str, strlen (buf->str));
 	g_string_free (buf, FALSE);
